refactor: no extra work for CSS unescaping
