+++ lib/dependencies/CssLocalIdentifierDependency.js
@@ -9,6 +9,7 @@ const { cssExportConvention } = require("../util/conventions");
 const createHash = require("../util/createHash");
 const { makePathsRelative } = require("../util/identifier");
 const makeSerializable = require("../util/makeSerializable");
+const memoize = require("../util/memoize");
 const NullDependency = require("./NullDependency");
 
 /** @typedef {import("webpack-sources").ReplaceSource} ReplaceSource */
@@ -29,6 +30,8 @@ const NullDependency = require("./NullDependency");
 /** @typedef {import("../serialization/ObjectMiddleware").ObjectSerializerContext} ObjectSerializerContext */
 /** @typedef {import("../util/Hash")} Hash */
 /** @typedef {import("../util/createHash").Algorithm} Algorithm */
+
+const getCssParser = memoize(() => require("../css/CssParser"));
 
 /**
  * @param {string} local css local
@@ -78,50 +81,6 @@ const getLocalIdent = (local, module, chunkGraph, runtimeTemplate) => {
 		.replace(/^((-?[0-9])|--)/, "_$1");
 };
 
-const CONTAINS_ESCAPE = /\\/;
-
-/**
- * @param {string} str string
- * @returns {[string, number] | undefined} hex
- */
-const gobbleHex = str => {
-	const lower = str.toLowerCase();
-	let hex = "";
-	let spaceTerminated = false;
-
-	for (let i = 0; i < 6 && lower[i] !== undefined; i++) {
-		const code = lower.charCodeAt(i);
-		// check to see if we are dealing with a valid hex char [a-f|0-9]
-		const valid = (code >= 97 && code <= 102) || (code >= 48 && code <= 57);
-		// https://drafts.csswg.org/css-syntax/#consume-escaped-code-point
-		spaceTerminated = code === 32;
-		if (!valid) break;
-		hex += lower[i];
-	}
-
-	if (hex.length === 0) return undefined;
-
-	const codePoint = Number.parseInt(hex, 16);
-	const isSurrogate = codePoint >= 0xd800 && codePoint <= 0xdfff;
-
-	// Add special case for
-	// "If this number is zero, or is for a surrogate, or is greater than the maximum allowed code point"
-	// https://drafts.csswg.org/css-syntax/#maximum-allowed-code-point
-	if (isSurrogate || codePoint === 0x0000 || codePoint > 0x10ffff) {
-		return ["\uFFFD", hex.length + (spaceTerminated ? 1 : 0)];
-	}
-
-	return [
-		String.fromCodePoint(codePoint),
-		hex.length + (spaceTerminated ? 1 : 0)
-	];
-};
-
-// eslint-disable-next-line no-useless-escape
-const regexSingleEscape = /[ -,.\/:-@[\]\^`{-~]/;
-const regexExcessiveSpaces =
-	/(^|\\+)?(\\[A-F0-9]{1,6})\u0020(?![a-fA-F0-9\u0020])/g;
-
 class CssLocalIdentifierDependency extends NullDependency {
 	/**
 	 * @param {string} name name
@@ -130,97 +89,11 @@ class CssLocalIdentifierDependency extends NullDependency {
 	 */
 	constructor(name, range, prefix = "") {
 		super();
-		this.name = CssLocalIdentifierDependency.unescapeIdentifier(name);
+		this.name = name;
 		this.range = range;
 		this.prefix = prefix;
 		this._conventionNames = undefined;
 		this._hashUpdate = undefined;
-	}
-
-	/**
-	 * @param {string} str string
-	 * @returns {string} unescaped string
-	 */
-	static unescapeIdentifier(str) {
-		const needToProcess = CONTAINS_ESCAPE.test(str);
-		if (!needToProcess) return str;
-		let ret = "";
-		for (let i = 0; i < str.length; i++) {
-			if (str[i] === "\\") {
-				const gobbled = gobbleHex(str.slice(i + 1, i + 7));
-				if (gobbled !== undefined) {
-					ret += gobbled[0];
-					i += gobbled[1];
-					continue;
-				}
-				// Retain a pair of \\ if double escaped `\\\\`
-				// https://github.com/postcss/postcss-selector-parser/commit/268c9a7656fb53f543dc620aa5b73a30ec3ff20e
-				if (str[i + 1] === "\\") {
-					ret += "\\";
-					i += 1;
-					continue;
-				}
-				// if \\ is at the end of the string retain it
-				// https://github.com/postcss/postcss-selector-parser/commit/01a6b346e3612ce1ab20219acc26abdc259ccefb
-				if (str.length === i + 1) {
-					ret += str[i];
-				}
-				continue;
-			}
-			ret += str[i];
-		}
-
-		return ret;
-	}
-
-	/**
-	 * @param {string} str string
-	 * @returns {string} escaped identifier
-	 */
-	static escapeIdentifier(str) {
-		let output = "";
-		let counter = 0;
-
-		while (counter < str.length) {
-			const character = str.charAt(counter++);
-
-			let value;
-
-			if (/[\t\n\f\r\u000B]/.test(character)) {
-				const codePoint = character.charCodeAt(0);
-
-				value = `\\${codePoint.toString(16).toUpperCase()} `;
-			} else if (character === "\\" || regexSingleEscape.test(character)) {
-				value = `\\${character}`;
-			} else {
-				value = character;
-			}
-
-			output += value;
-		}
-
-		const firstChar = str.charAt(0);
-
-		if (/^-[-\d]/.test(output)) {
-			output = `\\-${output.slice(1)}`;
-		} else if (/\d/.test(firstChar)) {
-			output = `\\3${firstChar} ${output.slice(1)}`;
-		}
-
-		// Remove spaces after `\HEX` escapes that are not followed by a hex digit,
-		// since they’re redundant. Note that this is only possible if the escape
-		// sequence isn’t preceded by an odd number of backslashes.
-		output = output.replace(regexExcessiveSpaces, ($0, $1, $2) => {
-			if ($1 && $1.length % 2) {
-				// It’s not safe to remove the space, so don’t.
-				return $0;
-			}
-
-			// Strip the space.
-			return ($1 || "") + $2;
-		});
-
-		return output;
 	}
 
 	get type() {
@@ -325,7 +198,7 @@ CssLocalIdentifierDependency.Template = class CssLocalIdentifierDependencyTempla
 
 		return (
 			dep.prefix +
-			CssLocalIdentifierDependency.escapeIdentifier(
+			getCssParser().escapeIdentifier(
 				getLocalIdent(local, module, chunkGraph, runtimeTemplate)
 			)
 		);
